requirements.txt
nicegui~=1.3.5
starlette~=0.27.0
ConfigArgParse~=1.2.3
gpiozero~=1.6.2
numpy~=1.20.0
paho-mqtt~=1.6.1
pyserial~=3.5
wifi~=0.3.8
fastapi~=0.100.0
requests~=2.25.1